refactor : 펫과 유저가 실제로 같은 그룹에 있는지 체크하는 로직에 필요한 로그인 유저 정보 aspect로 이동 #5
기존에는 @AuthenticationPrincipal를 통해 컨트롤러에서 로그인 유저 정보를 가져와서 서비스의 메서드에 전달했으나, 사용하지도 않는 이 userId를 굳이 전달할 필요가 없다고 생각.
